+++ yggdrasil/communication/ImageFileBase.py
@@ -89,7 +89,7 @@ class PILFileBase(DedicatedFileBase):
         return out
 
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Method to return a dictionary of testing options for this
         class.