+++ yggdrasil/communication/NetCDFFileComm.py
@@ -135,7 +135,7 @@ class NetCDFFileComm(DedicatedFileBase):
         return out
 
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Method to return a dictionary of testing options for this
         class.