+++ yggdrasil/communication/transforms/ArrayTransform.py
@@ -344,7 +344,7 @@ class ArrayTransform(TransformBase):
         return out
     
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Get testing options for the transform class.
 
         Returns: