+++ yggdrasil/communication/transforms/FilterTransform.py
@@ -56,7 +56,7 @@ class FilterTransform(TransformBase):
         return iter([])
 
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Get testing options for the transform class.
 
         Returns: