+++ yggdrasil/communication/transforms/SelectFieldsTransform.py
@@ -149,7 +149,7 @@ class SelectFieldsTransform(TransformBase):
         return out
     
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Get testing options for the transform class.
 
         Returns: