+++ yggdrasil/drivers/ModelDriver.py
@@ -4361,7 +4361,7 @@ class ModelDriver(Driver):
         return out
 
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Method to return a dictionary of testing options for this class.
 
         Returns: