+++ yggdrasil/serialize/CABOSerialize.py
@@ -186,7 +186,7 @@ class CABOSerialize(AsciiMapSerialize):
         return [total]
         
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Method to return a dictionary of testing options for this class.
 
         Returns: