+++ yggdrasil/serialize/MatSerialize.py
@@ -75,7 +75,7 @@ class MatSerialize(SerializeBase):
         return [total]
         
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Method to return a dictionary of testing options for this class.
 
         Returns: