+++ yggdrasil/serialize/PlySerialize.py
@@ -413,7 +413,7 @@ class PlySerialize(SerializeBase):
         return [out]
         
     @classmethod
-    def get_testing_options(cls):
+    def get_testing_options(cls, **kwargs):
         r"""Method to return a dictionary of testing options for this class.
 
         Returns: